requirements.txt
pyvisa
pyyaml
pyqtgraph
pyside6
